# Fixed memory issue