--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log
 
+### 0.2.7
+- Switch to `jsonschema_rs` for performance enhancement, instead of `jsonschema` package
+- Refactor code for improve memory utilization
+- Added garbage collector
+
+
 ### 0.2.6
 - Add garbage collection to free up memory after validation