--- a/src/python_osw_validation/extracted_data_validator.py
+++ b/src/python_osw_validation/extracted_data_validator.py
@@ -1,4 +1,5 @@
 import os
+import gc
 import glob
 
 OSW_DATASET_FILES = {
@@ -57,44 +58,55 @@ class ExtractedDataValidator:
         optional_files = [key for key, value in OSW_DATASET_FILES.items() if not value['required']]
         missing_files = []
         duplicate_files = []
-        for required_file in required_files:
-            file_count = 0
-            for filename in geojson_files:
-                base_name = os.path.basename(filename)
-                if required_file in base_name and base_name.endswith('.geojson'):
-                    file_count += 1
-                    save_filename = filename
-            if file_count == 0:
-                # Missing required file
-                missing_files.append(required_file)
-            elif file_count == 1:
-                self.files.append(save_filename)
-            else:
-                # Duplicate file
-                duplicate_files.append(required_file)
+        save_filename = None  # Initialize this variable
 
-        for optional_file in optional_files:
-            file_count = 0
-            for filename in geojson_files:
-                base_name = os.path.basename(filename)
-                if optional_file in base_name and base_name.endswith('.geojson'):
-                    file_count += 1
-                    save_filename = filename
-            if file_count == 1:
-                self.files.append(save_filename)
-            elif file_count > 1:
-                # Duplicate file
-                duplicate_files.append(optional_file)
+        try:
+            # Process required files
+            for required_file in required_files:
+                file_count = 0
+                for filename in geojson_files:
+                    base_name = os.path.basename(filename)
+                    if required_file in base_name and base_name.endswith('.geojson'):
+                        file_count += 1
+                        save_filename = filename
+                if file_count == 0:
+                    # Missing required file
+                    missing_files.append(required_file)
+                elif file_count == 1:
+                    self.files.append(save_filename)
+                else:
+                    # Duplicate file
+                    duplicate_files.append(required_file)
 
-        if missing_files:
-            self.error = f'Missing required .geojson files: {", ".join(missing_files)}.'
-            return False
+            # Process optional files
+            for optional_file in optional_files:
+                file_count = 0
+                for filename in geojson_files:
+                    base_name = os.path.basename(filename)
+                    if optional_file in base_name and base_name.endswith('.geojson'):
+                        file_count += 1
+                        save_filename = filename
+                if file_count == 1:
+                    self.files.append(save_filename)
+                elif file_count > 1:
+                    # Duplicate file
+                    duplicate_files.append(optional_file)
 
-        if duplicate_files:
-            self.error = f'Multiple .geojson files of the same type found: {", ".join(duplicate_files)}.'
-            return False
+            # Check for missing or duplicate files
+            if missing_files:
+                self.error = f'Missing required .geojson files: {", ".join(missing_files)}.'
+                return False
 
-        # Add OSW external extensions, GeoJSON files we know nothing about
-        self.externalExtensions.extend([item for item in geojson_files if item not in self.files])
+            if duplicate_files:
+                self.error = f'Multiple .geojson files of the same type found: {", ".join(duplicate_files)}.'
+                return False
+
+            # Add OSW external extensions, GeoJSON files we know nothing about
+            self.externalExtensions.extend([item for item in geojson_files if item not in self.files])
+
+        finally:
+            # Cleanup large lists and call garbage collector
+            del geojson_files, required_files, optional_files, missing_files, duplicate_files
+            gc.collect()
 
         return True