### Improvements Made:

- **Switch to jsonschema_rs** for performance enhancement, took from [PR #24](https://github.com/TaskarCenterAtUW/TDEI-python-lib-osw-validation/pull/24)
- **Refactored code** to improve memory utilization.
- **Integrated garbage collection** to release memory after processing large data.

### Other Changes:
- Updated package version
- Updated ChangeLog

### Sample Request Times

**Without Fix:**

- 10 MB file - 216.56 seconds
- 30 MB file - 292.48 seconds
- 50 MB file - 347.88 seconds
- 70 MB file - 338.3 seconds
- 90 MB file - 410.42 seconds
- 100 MB file - 518.55 seconds
- 200 MB file - 606.67 seconds
- 250 MB file - 573.07 seconds

**With Fix:**

- 10 MB file - 1.26 seconds
- 30 MB file - 3.84 seconds
- 50 MB file - 6.50 seconds
- 70 MB file - 9.17 seconds
- 90 MB file - 11.88 seconds
- 100 MB file - 13.27 seconds
- 200 MB file - 27.04 seconds
- 250 MB file - 33.67 seconds

### Percentage Improvement (Approximate):

- 10 MB file - **99.42%**
- 30 MB file - **98.69%**
- 50 MB file - **98.13%**
- 70 MB file - **97.29%**
- 90 MB file - **97.11%**
- 100 MB file - **97.44%**
- 200 MB file - **95.54%**
- 250 MB file - **94.13%**