@Stranger6667 2024-10-24
Hey, please, don’t hesitate to tag me if you’d anything from jsonschema-rs side or if it won’t run as fast as you’d expect it to run. I’d be eager to improve the library :)